+++ link/src/main/java/com/stripe/android/link/LinkActivity.kt
@@ -5,6 +5,7 @@ import android.os.Bundle
 import android.view.ViewTreeObserver
 import android.view.WindowManager
 import androidx.activity.ComponentActivity
+import androidx.activity.compose.BackHandler
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import androidx.annotation.VisibleForTesting
@@ -78,6 +79,7 @@ internal class LinkActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         // TODO(brnunes-stripe): Migrate to androidx.compose.foundation 1.2.0 when out of beta
         window.setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_ADJUST_RESIZE)
+        overridePendingTransition(R.anim.slide_up, 0)
 
         setContent {
             var bottomSheetContent by remember { mutableStateOf<BottomSheetContent?>(null) }
@@ -115,17 +117,19 @@ internal class LinkActivity : ComponentActivity() {
                     Column(Modifier.fillMaxWidth()) {
                         val linkAccount by viewModel.linkAccount.collectAsState(null)
                         val isOnRootScreen by isRootScreenFlow().collectAsState(true)
-
                         val backStackEntry by navController.currentBackStackEntryAsState()
                         val appBarState = rememberLinkAppBarState(
                             isRootScreen = isOnRootScreen,
                             currentRoute = backStackEntry?.destination?.route,
-                            email = linkAccount?.email
+                            email = linkAccount?.email,
+                            accountStatus = linkAccount?.accountStatus
                         )
+
+                        BackHandler(onBack = viewModel::onBackPressed)
 
                         LinkAppBar(
                             state = appBarState,
-                            onBackPressed = viewModel::onBackPressed,
+                            onBackPressed = onBackPressedDispatcher::onBackPressed,
                             onLogout = viewModel::logout,
                             showBottomSheetContent = {
                                 if (it == null) {
@@ -265,6 +269,11 @@ internal class LinkActivity : ComponentActivity() {
     override fun onDestroy() {
         super.onDestroy()
         viewModel.unregisterFromActivity()
+    }
+
+    override fun finish() {
+        super.finish()
+        overridePendingTransition(0, R.anim.slide_down)
     }
 
     private fun dismiss(result: LinkActivityResult) {