+++ link/src/main/java/com/stripe/android/link/ui/paymentmethod/PaymentMethodBody.kt
@@ -1,6 +1,7 @@
 package com.stripe.android.link.ui.paymentmethod
 
 import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
@@ -197,9 +198,9 @@ internal fun PaymentMethodBody(
             }
             Spacer(modifier = Modifier.height(8.dp))
         }
-        errorMessage?.let {
+        AnimatedVisibility(visible = errorMessage != null) {
             ErrorText(
-                text = it.getMessage(LocalContext.current.resources),
+                text = errorMessage?.getMessage(LocalContext.current.resources).orEmpty(),
                 modifier = Modifier.fillMaxWidth()
             )
         }