+++ payments-core/src/main/java/com/stripe/android/networking/StripeRepository.kt
@@ -10,6 +10,7 @@ import com.stripe.android.core.exception.InvalidRequestException
 import com.stripe.android.core.model.StripeFile
 import com.stripe.android.core.model.StripeFileParams
 import com.stripe.android.core.networking.ApiRequest
+import com.stripe.android.core.networking.StripeResponse
 import com.stripe.android.exception.CardException
 import com.stripe.android.model.BankStatuses
 import com.stripe.android.model.CardMetadata
@@ -40,7 +41,6 @@ import com.stripe.android.model.StripeIntent
 import com.stripe.android.model.Token
 import com.stripe.android.model.TokenParams
 import org.json.JSONException
-import org.json.JSONObject
 import java.util.Locale
 
 /**
@@ -391,7 +391,7 @@ abstract class StripeRepository {
     internal abstract suspend fun retrieveObject(
         url: String,
         requestOptions: ApiRequest.Options
-    ): JSONObject
+    ): StripeResponse<String>
 
     internal abstract suspend fun createRadarSession(
         requestOptions: ApiRequest.Options